add modal, put arrow and thubnails on top of img
