+++ src/components/Overview/Overview.jsx
@@ -12,8 +12,6 @@ const Overview = ({productId, productDetails}) => {
   const [styles, setStyles] = useState([])
   const [currentStyle, setCurrentStyle] = useState(styles[0])
 
-  console.log(currentStyle);
-
   const getStyles = (productId) => {
     axios.get('/getStyles', {params: {id: productId}})
     .then(response => {
