+++ src/components/Overview/ProductOverview.jsx
@@ -1,9 +1,10 @@
 import React, {useEffect, useState} from 'react';
+import { FaShare, FaTwitter, FaPinterest, FaFacebook } from 'react-icons/fa';
 
 const ProductOverview = ({productDetails}) => {
   return (
     <div className='ProductOverview'>
-      <h3>{productDetails.slogan}</h3>
+      <h3>{productDetails.slogan}&nbsp;&nbsp;<FaFacebook/>&nbsp;&nbsp;<FaTwitter/>&nbsp;&nbsp;<FaPinterest/></h3>
       <p>{productDetails.description}</p>
     </div>
   )
