--- test/OverviewTest.js
@@ -1,1 +1,0 @@
-import renderer from 'react-test-renderer';
