--- test/RelatedTest.js
@@ -1,2 +1,0 @@
-import renderer from 'react-test-renderer';
-import ProductCardRI from '../src/components/Related/ProductCardRI.jsx';
